Games and stats rendering, main part ready

+++ src/SharedServices/JsonMessages/RequestPlayersTopMessage.cs
@@ -2,7 +2,7 @@
 {
     public class RequestPlayersTopMessage
     {
-        public long Offset { get; set; }
-        public long Count { get; set; }
+        public int Offset { get; set; }
+        public int Count { get; set; }
     }
 }
